+++ richat/src/bin/richat.rs
@@ -37,7 +37,7 @@ fn main() -> anyhow::Result<()> {
     }
 
     // Setup logs
-    richat::log::setup(config.log.json)?;
+    richat::log::setup(config.logs.json)?;
 
     // Shutdown channel/flag
     let shutdown = Shutdown::new();
@@ -129,7 +129,7 @@ fn main() -> anyhow::Result<()> {
                     ready(Ok(())).boxed()
                 };
 
-                let prometheus_fut = if let Some(config) = config.prometheus {
+                let metrics_fut = if let Some(config) = config.metrics {
                     richat::metrics::spawn_server(config, shutdown)
                         .await?
                         .map_err(anyhow::Error::from)
@@ -138,9 +138,7 @@ fn main() -> anyhow::Result<()> {
                     ready(Ok(())).boxed()
                 };
 
-                try_join_all(vec![grpc_fut, prometheus_fut])
-                    .await
-                    .map(|_| ())
+                try_join_all(vec![grpc_fut, metrics_fut]).await.map(|_| ())
             })
         }
     })?;