+++ shared/src/config.rs
@@ -55,12 +55,12 @@ impl ConfigTokio {
 
 #[derive(Debug, Clone, Copy, Deserialize)]
 #[serde(deny_unknown_fields, default)]
-pub struct ConfigPrometheus {
+pub struct ConfigMetrics {
     /// Endpoint of Prometheus service
     pub endpoint: SocketAddr,
 }
 
-impl Default for ConfigPrometheus {
+impl Default for ConfigMetrics {
     fn default() -> Self {
         Self {
             endpoint: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 10123),